♻️ テナント登録処理の分離
