+++ packages/cdk/lib/stacks/tenant/tenant-iam-stack.ts
@@ -2,8 +2,6 @@ import * as cdk from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { UserPool, IUserPool } from 'aws-cdk-lib/aws-cognito';
 import { IdentityPool, IIdentityPool } from 'aws-cdk-lib/aws-cognito-identitypool';
-import { Runtime, Code } from 'aws-cdk-lib/aws-lambda';
-import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
 import { TenantRole } from '../../construct/tenant-role';
 
 export interface TenantIAMStackProps extends cdk.StackProps {
@@ -75,20 +73,6 @@ export class TenantIAMStack extends cdk.Stack {
       );
     }
 
-    // Get tenant registration API endpoint and key from context
-    const registrationApiEndpoint = this.node.tryGetContext('registrationApiEndpoint');
-    const registrationApiKey = this.node.tryGetContext('registrationApiKey');
-    if (!registrationApiEndpoint) {
-      throw new Error(
-        'registrationApiEndpoint must be provided via context (--context registrationApiEndpoint=<value> or in cdk.tenant.json)'
-      );
-    }
-    if (!registrationApiKey) {
-      throw new Error(
-        'registrationApiKey must be provided via context (--context registrationApiKey=<value> or in cdk.tenant.json)'
-      );
-    }
-
     // Import existing pools using the context values from main stack
     const userPool = UserPool.fromUserPoolId(this, 'ImportedUserPool', userPoolId);
     const identityPool = IdentityPool.fromIdentityPoolId(this, 'ImportedIdentityPool', identityPoolId);
@@ -103,153 +87,6 @@ export class TenantIAMStack extends cdk.Stack {
       account: this.account,
       env: environment,
     });
-
-    // Create a Lambda to call the registration API
-    const registerTenantLambda = new NodejsFunction(this, 'RegisterTenant', {
-      functionName: `tenant-registration-caller-${this.tenantId}`,
-      runtime: Runtime.NODEJS_18_X,
-      handler: 'index.handler',
-      timeout: cdk.Duration.seconds(30),
-      code: Code.fromInline(`
-        const https = require('https');
-        const { URL } = require('url');
-        
-        // Send response to CloudFormation
-        const sendResponse = async (event, status, reason, physicalResourceId, data = {}) => {
-          const responseBody = JSON.stringify({
-            Status: status,
-            Reason: reason,
-            PhysicalResourceId: physicalResourceId,
-            StackId: event.StackId,
-            RequestId: event.RequestId,
-            LogicalResourceId: event.LogicalResourceId,
-            NoEcho: false,
-            Data: data,
-          });
-          
-          console.log('Sending CloudFormation response:', responseBody);
-          
-          const responseUrl = new URL(event.ResponseURL);
-          const options = {
-            hostname: responseUrl.hostname,
-            port: 443,
-            path: responseUrl.pathname + responseUrl.search,
-            method: 'PUT',
-            headers: {
-              'Content-Type': '',
-              'Content-Length': responseBody.length,
-            },
-          };
-          
-          return new Promise((resolve, reject) => {
-            const req = https.request(options, (res) => {
-              console.log('CloudFormation response status:', res.statusCode);
-              resolve();
-            });
-            
-            req.on('error', (error) => {
-              console.error('Failed to send CloudFormation response:', error);
-              reject(error);
-            });
-            
-            req.write(responseBody);
-            req.end();
-          });
-        };
-        
-        exports.handler = async (event, context) => {
-          console.log('Event:', JSON.stringify(event, null, 2));
-          
-          const physicalResourceId = event.PhysicalResourceId || 'tenant-registration-${this.tenantId}';
-          
-          try {
-            if (event.RequestType === 'Delete') {
-              // For deletion, we don't need to call the API
-              await sendResponse(event, 'SUCCESS', 'Delete completed successfully', physicalResourceId);
-              return;
-            }
-            
-            if (event.RequestType === 'Update') {
-              // For updates, we could re-register or just return success
-              await sendResponse(event, 'SUCCESS', 'Update completed successfully', physicalResourceId);
-              return;
-            }
-            
-            // Handle Create request
-            const endpoint = '${registrationApiEndpoint}';
-            const apiKey = '${registrationApiKey}';
-            
-            const data = JSON.stringify({
-              tenantId: '${this.tenantId}',
-              accountId: '${this.account}',
-              region: '${this.region}',
-              environment: '${environment}',
-              roleArn: '${this.tenantRole.role.roleArn}',
-            });
-            
-            console.log('Calling registration API:', endpoint);
-            console.log('Registration data:', data);
-            
-            const url = new URL(endpoint);
-            const options = {
-              hostname: url.hostname,
-              port: 443,
-              path: url.pathname,
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/json',
-                'Content-Length': data.length,
-                'x-api-key': apiKey,
-              },
-            };
-            
-            await new Promise((resolve, reject) => {
-              const req = https.request(options, (res) => {
-                let body = '';
-                res.on('data', (chunk) => body += chunk);
-                res.on('end', async () => {
-                  console.log('API Response Status:', res.statusCode);
-                  console.log('API Response Body:', body);
-                  
-                  if (res.statusCode === 200 || res.statusCode === 409) {
-                    // 200 = success, 409 = already exists (acceptable)
-                    await sendResponse(event, 'SUCCESS', 'Tenant registered successfully', physicalResourceId, { registered: true });
-                    resolve();
-                  } else {
-                    const errorMsg = 'Registration API call failed with status ' + res.statusCode + ': ' + body;
-                    await sendResponse(event, 'FAILED', errorMsg, physicalResourceId);
-                    reject(new Error(errorMsg));
-                  }
-                });
-              });
-              
-              req.on('error', async (error) => {
-                console.error('Request error:', error);
-                await sendResponse(event, 'FAILED', 'Request error: ' + error.message, physicalResourceId);
-                reject(error);
-              });
-              
-              req.write(data);
-              req.end();
-            });
-            
-          } catch (error) {
-            console.error('Lambda execution error:', error);
-            await sendResponse(event, 'FAILED', 'Lambda execution error: ' + error.message, physicalResourceId);
-            throw error;
-          }
-        };
-      `),
-    });
-
-    // Create Custom Resource using the local Lambda
-    const tenantRegistrationResource = new cdk.CustomResource(this, 'TenantRegistration', {
-      serviceToken: registerTenantLambda.functionArn,
-      resourceType: 'Custom::TenantRegistration',
-    });
-
-    // Ensure registration happens after role creation
-    tenantRegistrationResource.node.addDependency(this.tenantRole);
 
     // Add stack-level outputs with export names
     new cdk.CfnOutput(this, 'StackTenantRoleArn', {
