+++ packages/cdk/lib/types/tenant-registration.ts
@@ -1,0 +1,18 @@
+export interface TenantRegistrationData {
+  tenantId: string;
+  accountId: string;
+  region: string;
+  environment: string;
+  roleArn?: string;
+  bedrockChatApiArn?: string;
+}
+
+export interface TenantRegistrationRequest extends TenantRegistrationData {
+}
+
+export enum TenantStatus {
+  ACTIVE = 'active',
+  INACTIVE = 'inactive',
+  PROVISIONING = 'provisioning',
+  ERROR = 'error',
+}
